+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Bug Fix
 - Fixed a bug in Bootstrap styling that caused validation errors to force input fields downwards
 - Fixed a bug that caused the `next` and `prev` buttons on reports to not use the correct report filters
+- Fixed a bug that would overwrite the original creator of a comment, contact, document, domain, link, and post when updated by another user
 
 ## 0.15.1 (April 3, 2013)